awsutils: move package into ./pkg
