+++ input/sqs.go
@@ -11,8 +11,8 @@ import (
 	log "github.com/sirupsen/logrus"
 
 	"github.com/AdRoll/baker"
-	"github.com/AdRoll/baker/awsutils"
 	"github.com/AdRoll/baker/input/inpututils"
+	"github.com/AdRoll/baker/pkg/awsutils"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/sqs"
