+++ output/dyndb.go
@@ -7,13 +7,14 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/AdRoll/baker"
-	"github.com/AdRoll/baker/awsutils"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 	"github.com/juju/ratelimit"
 	log "github.com/sirupsen/logrus"
+
+	"github.com/AdRoll/baker"
+	"github.com/AdRoll/baker/pkg/awsutils"
 )
 
 var DynamoDBDesc = baker.OutputDesc{
